scraper/requirements.txt
supabase==2.3.4
gotrue==2.2.2
httpx==0.27.0
geopy==2.4.1
python-dotenv==1.0.1